fix(curriculum): single source of truth for topic order + migration timeouts
Two follow-ups to PR #473 surfaced during review:

1. Topic order had two sources of truth in YAML -- the slug list in
   _phase.yaml AND an 'order' field inside each topic file. They could
   drift silently because the loader read the explicit field but the UI
   rendered from the slug list. Now the slug list is canonical; the
   loader injects 'order' from the position. Topic YAML files no longer
   carry 'order' -- the loader rejects it loudly if reintroduced.

2. Migration 0025 sets lock_timeout=5s and statement_timeout=30s at
   the top of the transaction so a stalled lock or runaway query can't
   hang the deploy (squawk's recommended pattern). Brand-new empty
   tables run instantly under these limits.

Also: legacy_id and order columns use BigInteger to satisfy squawk's
prefer-bigint-over-int rule. Properly fixed rather than excluding the
rule -- INTEGER caps at 2.1B which is fine for our data sizes, but
BIGINT costs 4 extra bytes per row for permanent overflow safety on
new schemas.

Implementation:
* schemas.Topic still declares order: int (loader-populated, not from YAML)
* _load_topic gains an 'order' kwarg supplied by _load_phase based on
  slug-list position (idx + 1)
* _load_topic raises ContentValidationError if topic YAML contains
  'order' so future authoring mistakes fail loudly
* scripts/strip_topic_order.py: one-time idempotent script that removed
  the 'order' field from all 42 topic YAML files
* Updated test_content_service.py: TestLoadTopic now passes order=
  parameter; new test_topic_with_order_field_is_rejected guards the
  invariant
* Added a 'No Hacks or Bandaids' section to
  .github/copilot-instructions.md codifying the user preference that
  drove fix 1 (don't paper over warnings; do the real refactor)

Quality gates: ruff + format + ty + pytest (226 API, 449 shared, +1
new test) all green. Squawk on the new migration now clean. End-to-end
sync against real local DB: 7 phases, 42 topics, 272 steps, 135
objectives, 10 requirements -- counts unchanged, confirming no
regression.

+++ packages/learn-to-cloud-shared/src/learn_to_cloud_shared/models.py
@@ -350,12 +350,12 @@ class CurriculumPhase(TimestampMixin, Base):
     )
 
     uuid: Mapped[UUID] = mapped_column(Uuid(as_uuid=True), primary_key=True)
-    legacy_id: Mapped[int] = mapped_column(Integer, nullable=False)
+    legacy_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     slug: Mapped[str] = mapped_column(Text, nullable=False)
     name: Mapped[str] = mapped_column(Text, nullable=False)
     description: Mapped[str] = mapped_column(Text, nullable=False)
     short_description: Mapped[str] = mapped_column(Text, nullable=False)
-    order: Mapped[int] = mapped_column(Integer, nullable=False)
+    order: Mapped[int] = mapped_column(BigInteger, nullable=False)
     deleted_at: Mapped[datetime | None] = mapped_column(
         DateTime(timezone=True), nullable=True
     )
@@ -385,7 +385,7 @@ class CurriculumTopic(TimestampMixin, Base):
     slug: Mapped[str] = mapped_column(Text, nullable=False)
     name: Mapped[str] = mapped_column(Text, nullable=False)
     description: Mapped[str] = mapped_column(Text, nullable=False)
-    order: Mapped[int] = mapped_column(Integer, nullable=False)
+    order: Mapped[int] = mapped_column(BigInteger, nullable=False)
     deleted_at: Mapped[datetime | None] = mapped_column(
         DateTime(timezone=True), nullable=True
     )
@@ -412,7 +412,7 @@ class CurriculumStep(TimestampMixin, Base):
         nullable=False,
     )
     legacy_id: Mapped[str] = mapped_column(Text, nullable=False)
-    order: Mapped[int] = mapped_column(Integer, nullable=False)
+    order: Mapped[int] = mapped_column(BigInteger, nullable=False)
     action: Mapped[str | None] = mapped_column(Text, nullable=True)
     title: Mapped[str | None] = mapped_column(Text, nullable=True)
     url: Mapped[str | None] = mapped_column(Text, nullable=True)
@@ -453,7 +453,7 @@ class CurriculumLearningObjective(TimestampMixin, Base):
     )
     legacy_id: Mapped[str] = mapped_column(Text, nullable=False)
     text_: Mapped[str] = mapped_column("text", Text, nullable=False)
-    order: Mapped[int] = mapped_column(Integer, nullable=False)
+    order: Mapped[int] = mapped_column(BigInteger, nullable=False)
     deleted_at: Mapped[datetime | None] = mapped_column(
         DateTime(timezone=True), nullable=True
     )
